fix: reject --probe-matrix/--post-manifest and fix scope-block gap for --old-bundle-facts
Codex review, fresh evidence: neither --probe-matrix nor --post-manifest
had a channel into compare_release_against_bundle_facts() or its
per-library service.compare_snapshots() call, so both were silently
accepted and ignored -- now rejected explicitly (exit 64), matching the
existing "reject rather than silently diverge" rule this dispatcher
already applies to every other unsupported flag. The project-config
scope: block rejection also missed two of its three independent fields
(collapse_versioned_symbols, public_symbols) -- a config setting only
those, with scope_public left None, previously passed the check
unrejected even though neither is applied in this mode either.

Split TestCompareOldBundleFactsEarlyRejections out of
test_cli_compare_bundle_facts.py into its own sibling file (matching
the earlier test_run_plan_public_headers.py split) after the added
regression tests pushed the original past the architecture no-growth
test-file cap.
